Move live viz example
