+++ python/foxglove-sdk-examples/live_visualization/pyproject.toml
@@ -1,0 +1,18 @@
+[project]
+name = "live-visualization"
+version = "0.0.0"
+description = "Example of live visualization in Foxglove"
+authors = [{ name = "Foxglove", email = "support@foxglove.dev" }]
+readme = "README.md"
+requires-python = ">=3.9"
+dependencies = ["foxglove-sdk @ file:../../foxglove-sdk/", "numpy@^2.0"]
+
+[build-system]
+requires = ["poetry-core>=2.0.0,<3.0.0"]
+build-backend = "poetry.core.masonry.api"
+
+[tool.poetry]
+package-mode = false
+
+[tool.poetry.group.dev.dependencies]
+mypy = "^1"
