--- python/foxglove-sdk/python/examples/geometry/__init__.py
@@ -1,5 +1,0 @@
-from .euler_to_quaternion import euler_to_quaternion
-
-__all__ = [
-    "euler_to_quaternion",
-]
